scheta_add2

+++ shpagin_podolsky/dictonary/view.php
@@ -9,22 +9,31 @@
         $curent_table = isset($_GET['table']) && in_array($_GET['table'], DICTONARY_TABLES)
                         ? $_GET['table'] : "s_accoount";
 
-        $result = $mysqli->query("SELECT * FROM $curent_table");
+        $result = $mysqli->query(
+            "SELECT
+                `".DICTONARY_TABLES_STRUCT[$curent_table]["id"]."`,
+                `".DICTONARY_TABLES_STRUCT[$curent_table]["name"]."`
+                
+            FROM $curent_table
+            ");
 
         $tableRows = [];
-        while($row = $result->fetch_assoc()){
-            $row = array_values($row);
+        while($row = $result->fetch_row()){
+            $subResult = $mysqli -> query("SELECT count(*) FROM `".DICTONARY_TABLES_STRUCT[$curent_table]["ref_table"]."` WHERE `".DICTONARY_TABLES_STRUCT[$curent_table]["id"]."` = {$row[0]}");
+            $count = $subResult -> fetch_row()[0];
+
             $tableRows[] = [
                 "id" => $row[0],
-                "name" => $row[1]
+                "name" => $row[1],
+                "access_delete" => $count == 0
             ];
         }
         
     ?>
     <div class="container text-center">
     <div class="row">
-    <div class="col-12 col-sm-12 col-md-12 col-lg-6 col-xl-4">
-        <div class="d-flex flex-column flex-shrink-0 p-3 text-black bg-white" style="width: 280px;">
+    <div class="col-12 col-lg-6 col-xl-4">
+        <div class="d-flex flex-column flex-shrink-0 p-3 text-black bg-white">
             <span class="fs-4">Список справочников</span>
             <hr>
                 <ul class="nav nav-pills flex-column mb-auto">
@@ -55,8 +64,8 @@
             <tr>
                 <td width="65%"><?= $row["name"] ?></td>
                 <td class="btns"> 
-                    <a type="button" class="btn btn-secondary">Редактировать</a> 
-                    <a type="button" class="btn btn-secondary" href="/shpagin_podolsky/dictonary/delete.php?table=<?=$curent_table?>&id=<?=$row["id"]?>">Удалить</a>
+                    <a type="button" class="btn btn-secondary" href="/shpagin_podolsky/dictonary/edit.php?table=<?=$curent_table?>&id=<?=$row["id"]?>">Редактировать</a> 
+                    <a type="button" class="btn btn-secondary <?= !$row["access_delete"] ? "disabled" : ""?>" href="/shpagin_podolsky/dictonary/delete.php?table=<?=$curent_table?>&id=<?=$row["id"]?>" onclick='return confirm(`Вы действительно хотите удалить запись "<?=$row["name"]?>"?`);'>Удалить</a>
                 </td>
             </tr>
             <?php } ?>
